CA-391 Check @Secured for session validation
- Adds AuthenticationFilter which responds to each REST call by virtue
of extending ContainerRequestFilter.
- Adds @Secured qualifier annotation for protected REST resources.
- Adds annotation to the MemberWebService to protect it.
- Adds interface and empty implementation for the AccessTokenService
responsible for validating the AccessToken against the 3rd-party auth
provider.

NOTE: There were some interesting NPEs until I was able to setup a "bean-style"
constructor; not able to use constructor-based injection for the AuthenticationFilter.
Further, none of the injected services are available at the time of construction.
Logging was a notable service.

Logging is also settling out within this commit:
- Log4j.properties as part of src/main/resources.
- Changed JUL over to SLF4J when generating the Logger resource for injection.
- Earlier commit placed the dependencies in the pom.xml.