Code supporting the GeoJson file format.
Came over from the old `clueride-angular` project that used JSON-based GeoJson files. This is able to parse the files and represent them in `Geometry` aware classes. However, this Geometry is not the same as the Geolatte geometry we're moving toward.

Also, this comes short of persisting because we don't have a JPA
`@PersistenceContext` that runs under SE yet. That will be in a separate ticket.