SVR-130 REST API for path/meta call can get the attraction list all messed up
- Changes approach for persisting Paths for a given set of Attractions and the links to those Paths from a Course.

This simplifies how the records are managed: Paths are created for pairs of Attractions just so the presence or absence of Edge
records can be checked. Once the Course is saved, the resulting list of settled Paths is then linked to the Course by
clearing out existing Link records and replacing with a new full set.

Adding constraint criteria to the DB records greatly helped in narrowing down where the issues were arising in the code.